fix(cli): Windows paths on separate drive (#316)
When building a Celest project from a drive other than `C:\`, the path
handling was previously broken because we were trying to make relative
paths which are not possible on Windows between drive letters.

Here, we still have pretty relative paths, but only when we can
guarantee the output is on the same drive letter as the imported file.