fix(cli): Feature-dependent codegen (#814)
Fixes: https://github.com/celest-dev/celest/issues/151
Fixes: https://celest.sentry.io/issues/5499486302/events/276cbd980f4a4a96a42b755ac8ccce78/
Fixes: https://celest.sentry.io/issues/5499486389/events/02e585e782f148e1bdf90fd88d321760/

- Adds the notion of feature flags and performs code-generation on that basis.
- Adds tests that confirm support of the `supabase` package.

Remaining TODOs:

This issue arose because `supabase` depends on `websocket_channel: ^2` while the latest version of the runtime
requires `websocket_channel: ^3`. Had the breaking change been released as a breaking version (e.g. 0.5.0), that
would have caught the error sooner, e.g. `pub add` would display a version conflict, but it would still prevent
users from adding supabase to their backend.

There are no great answers to this except dramatically limiting the number of dependencies that `celest` uses.
To do this, though, we'll need to have separate packages for the backend definition components and the generated
client. Further, we'll need to generate a separate Dart/Flutter package for the generated client so that there is
no overlap between the backend dependencies and the frontend dependencies.

This is a good initiative regardless, and will unblock other issues such as https://github.com/celest-dev/celest/issues/47.
This work will be scoped for 0.5.0.