chore(cli): Improve Flutter support (#1033)
There are a lot of nuances to supporting Flutter widgets over-the-wire. For now, we can at least improve the experience so that it's at least possible to
do it. Most of the gotchas can only be resolved with full bytecode support, which is still in its infancy, but the basic support available is enough to have
a cool demo of the future possibilities.