chore(cli): Prepare for publishing
The CLI was previously published by building an AOT executable and bundling with platform-dependent methods. This was nice for improving performance
but suffers from requiring a lot of infrastructure and having a non-zero cost.

With the open-sourcing of the CLI, we'll publish to pub.dev, sacrificing some cost but making installation a single `dart pub global activate` command.

- Removes old release code
- Updates `pubspec.yaml` to support publishing to pub.dev