chore(cli): Pass Celest's VM service URI to auxiliary start commands (#270)
When running `celest start` with an auxiliary command, pass the running server's VM service URI to the command as a Dart define.

This is useful when the auxiliary command is `dart test` or `flutter test` and allows `package:celest_test` to get live updates from the server in an integration test.