feat(cloud_auth): Make database modular (#289)
Allows the `CloudAuthDatabase` to be incorporated into other Drift databases via the `include` field and a `CloudAuthDatabaseMixin` mixin. This means that only one database class is required per project.

In order to control schema upgrades in the Cloud Auth portion of the database, a new meta table is created so that the version pragma can be used for the host database's schema version.